Move /cli and its subpackages to /internal
